Fix cache

+++ src/websocket/schema/Request/requestTransform.model.ts
@@ -2,7 +2,7 @@ import { z } from "zod";
 
 export const requestTransformSchema = z.object({
   duration_s: z.number().int().positive(),
-  frequency: z.number().int().positive(),
+  frequency: z.number().positive(),
   from_timestamp: z.number().int().min(0),
 });
